addressing code review comments from @joehan
