+++ scripts/emulator-tests/functions/index.js
@@ -1,0 +1,12 @@
+module.exports = (() => {
+                return {
+                    functionId: require("firebase-functions").https.onRequest((req, resp) => {
+                        return new Promise((resolve) => {
+                            setTimeout(() => {
+                                resp.sendStatus(200);
+                                resolve();
+                            }, 3000);
+                        });
+                    }),
+                };
+            })();
